assistant/requirements.txt
slack-bolt>=1.22.0
slack-sdk>=3.35.0
anthropic>=0.45.0
python-dotenv>=1.0.0
